program that prints the first 50 Fibonacci numbers,

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,7 +9,7 @@ int main(void)
 {
 	long i = 0;
 	long j = 1;
-	long n = 0;
+	int n = 0;
 	long fibonacci;
 	
 	while (n < 50)
